requirements.txt
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain
langchain_experimental